embedding with openai for lightgbm model, creating separate cleaner python file for embeddings
