+++ requirements.txt
@@ -12,3 +12,5 @@ gensim~=4.3.2
 scikit-learn~=1.2.1
 scipy~=1.10.0
 neattext~=0.1.3
+lightgbm
+openai
